pyproject.toml
[project]
name = "sdapi-discord"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "sserada", email = "hirawatasou@gmail.com" }
]
dependencies = [
    "discord-py>=2.5.2",
    "python-dotenv>=1.1.1",
    "httpx>=0.28.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/sdapi_discord"]

[tool.rye.scripts]
start = { cmd = "python -m src.sdapi_discord.main" }
